requirements.txt
asgiref==3.6.0
asttokens==2.2.1
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
black==23.1.0
blinker==1.5
certifi==2022.12.7
click==8.1.3
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
elasticsearch==7.12.1
email-validator==1.3.1
executing==1.2.0
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.0
Flask-WTF==1.0.0
idna==3.4
importlib-metadata==6.0.0
ipdb==0.13.11
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mongoengine==0.26.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
speaklater==1.3
sqlparse==0.4.3
stack-data==0.6.2
tdqm==0.0.1
tomli==2.0.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
